requirements.txt
transformers
torch
pymupdf
openai
python-pptx
llama-cpp-python
argparse
